made drawer easier to implement in an activity
